+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog since 1.7.2:
  - Add experimental support for the Corsair Hydro H60i Pro XT
 ### Fixed
  - Read DDR4 temperature sensor by word instead of with SMBus Block Read (#400)
+ - Fix tolerant handling of single channel name in Corsair Lighting Node Core
 
 
 ## [1.7.2] – 2021-10-05